Update configuring-the-tenant-loading-policy.md
When we set tenant_idle_timeout='30m', as 'm' is a character, the server starts with an error -  java.lang.NumberFormatException: For input string "30m". But when we set tenant_idle_timeout='30', the server starts properly without any error. So updating tenant_idle_timeout='30m' to tenant_idle_timeout='30' in the documentation.