Update getting-started-with-wso2-api-controller.md
The command "./apictl" to start the CTL tool doesn't work and gives out an error as " '.' is not recognized as an internal or external command,
operable program or batch file. ". Instead just running the command as "apictl" works and the directory structure for the configuration files ( <USER_HOME>/.wso2apictl ) gets created.